+++ src/config.ts
@@ -7,9 +7,11 @@ export function get(key: "clearConsoleOnRun"): boolean;
 export function get(
   key: "layout"
 ): "grid" | "splitLeft" | "splitRight" | "splitTop";
+export function get(key: "launchBehavior"): "newSwing" | "none" | "openSwing";
 export function get(
   key: "readmeBehavior"
 ): "none" | "previewFooter" | "previewHeader";
+export function get(key: "rootDirectory"): string;
 export function get(key: "showConsole"): boolean;
 export function get(key: "templateGalleries"): string[];
 export function get(key: "themePreview"): boolean;