+++ src/extension.ts
@@ -3,8 +3,8 @@ import { api } from "./api";
 import { registerCreationModule } from "./creation";
 import { registerLiveShareModule } from "./liveShare";
 import { registerPreviewModule } from "./preview";
+import { registerTreeViewModule } from "./preview/tree";
 import { store } from "./store";
-import { registerTreeViewModule } from "./tree";
 import { checkForSwingWorkspace } from "./utils";
 
 export async function activate(context: vscode.ExtensionContext) {