+++ src/utils.ts
@@ -1,5 +1,6 @@
-import { ProgressLocation, Uri, window, workspace } from "vscode";
-import { SWING_FILE } from "./constants";
+import { commands, ProgressLocation, Uri, window, workspace } from "vscode";
+import * as config from "./config";
+import { EXTENSION_NAME, SWING_FILE } from "./constants";
 import { openSwing } from "./preview";
 
 export function byteArrayToString(value: Uint8Array) {
@@ -11,12 +12,21 @@ export function stringToByteArray(value: string) {
 }
 
 export async function checkForSwingWorkspace() {
-  if (workspace.workspaceFolders) {
-    // TODO: Check how to make this work
-    // for non-file system workspaces
-    const files = await workspace.findFiles(SWING_FILE);
-    if (files.length > 0) {
-      openSwing(workspace.workspaceFolders[0].uri);
+  switch (config.get("launchBehavior")) {
+    case "openSwing": {
+      if (workspace.workspaceFolders) {
+        const files = await workspace.findFiles(SWING_FILE);
+        if (files.length > 0) {
+          openSwing(workspace.workspaceFolders[0].uri);
+        } else if (config.get("rootDirectory")) {
+          commands.executeCommand(`${EXTENSION_NAME}.initializeWorkspace`);
+        }
+      }
+      break;
+    }
+    case "newSwing": {
+      commands.executeCommand(`${EXTENSION_NAME}.newSwing`);
+      break;
     }
   }
 }