+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Coming Up
 
 - Fix CodePen exporting
+- Adding support for `window.open` with swing-local HTML files
+- Enabling SCSS/Sass support in github.dev
 - Grouping tutorial steps into sections (https://github.com/sveltejs/svelte/tree/master/site/content/examples)
+
+# v0.0.22 (09/19/21)
+
+- Introduced a new `CodeSwing: Launch Behavior` setting, that allows customizing how CodeSwing should behave when you open a swing workspace.
+- Introduced a new `CodeSwing: Root Directory` setting, that allows specifying the workspace directory that new swings should be created in
+- Introduced a new `CodeSwing: Initialize Workspace as Swing` command, that allows you to open a folder and turn it into a swing
+- Added support for calling `window.open()` on HTTP(S) URLs
 
 # v0.0.21 (07/26/21)